support .anki21 files
Newer versions of Anki save decks as `.anki21` instead of `.anki2`. This
change adds support for importing these decks.

Even newer versions of Anki save files as `.anki21b`, but the format of
these appears to be different. For now, notify the user about the
problem and tell them to use the export option that creates an
`.anki21` instead.